Don't drop Keypresses outside of readchar_linux()
Calling `tty.setraw` makes it so that subsequent calls to `read()` don't read
any of the typed-but-not-read characters. Reading the code, this could be
expected since `tty.setraw` is called without explicit parameters, and the
[default for when](https://github.com/python/cpython/blob/345572a1a026/Lib/tty.py#L18)
is `TCSAFLUSH`, which means:

> If optional_actions is TCSAFLUSH, the change shall occur after all output
> written to fildes is transmitted, **and all input so far received but not
> read shall be discarded before the change is made.** (emphasis mine)

Indeed, changing the call to be with `TCSADRAIN` seems to solve the problem of
missing keystrokes.

However, that solves the problem that any keystrokes that occur outside of
`readchar` are dropped fully, but it does not solve the problem that they will
be echoed. (They will be echoed because they occur at a point in time that
stdin is in "cooked" or "canonical" (non-raw) mode).

See #73