workflows: auto publish to pypi for every release
the new workflow changes the sourcecode to actually contain the correct
verison number befor building the distros. This ensures clean metadata.

The workflow also auto-increments the version number and commits the
new version to the repo. This means that the repos version is alway the
next developmentversion from the most recent release.